Make subspace source portable to QNX and add toolchain skeleton
Source-level changes so the C++ side of subspace can compile on QNX
without affecting Linux or macOS behaviour:

  * client/client.cc: GetThreadId() used reinterpret_cast<uint64_t> on
    pthread_self(), which only happens to work because pthread_t is
    pointer-sized on glibc and a pointer on macOS.  On QNX pthread_t is
    a 32-bit integer, where reinterpret_cast between unrelated integer
    types is ill-formed.  Use memcpy so the conversion is well-defined
    regardless of the underlying pthread_t representation.

  * common/channel.cc: PosixSharedMemoryName() formatted st_ino with
    "%d", which truncates 64-bit inodes (real on QNX and various Linux
    filesystems) and is technically the wrong type.  Promote to
    unsigned long long and use %llu.

  * server/main.cc: fix the long-standing `___APPLE__` triple-underscore
    typo that meant macOS hosts have always silently been getting the
    "non-Mac" default for --cleanup_filesystem.  While here, also turn
    the flag default on for QNX, where the server uses /tmp-shadowed
    POSIX shared memory just like macOS.

Bazel toolchain plumbing under bazel/toolchains/qnx/:

  * platforms/BUILD.bazel declares a `qnx` constraint_value attached to
    @platforms//os:os and the qnx_aarch64 / qnx_x86_64 platforms that
    .bazelrc has been pointing at.  This lets `--platforms=...` resolve
    to a real label.

  * BUILD.bazel and cc_toolchain_config.bzl provide a stub
    cc_toolchain wrapping qcc.  This is intentionally a skeleton: it
    declares the targets that downstream rules expect to find but
    relies on a real QNX SDP repository rule (not yet added) to wire
    the qcc/qld/q++ binaries and target sysroot in.  Builds with
    `--config=qnx_aarch64` will fail at the toolchain level until that
    is added; the source changes above are all reviewable independently.

Made-with: Cursor

+++ common/channel.cc
@@ -335,7 +335,10 @@ Channel::PosixSharedMemoryName(const std::string &shadow_file) {
                         shadow_file, strerror(errno)));
   }
   // Use the inode number (unique per file) to make the shm file name.
-  return absl::StrFormat("subspace_%d", st.st_ino);
+  // st_ino is unsigned and can be 64-bit on some systems (e.g. QNX), so
+  // promote to uint64_t before formatting.
+  return absl::StrFormat("subspace_%llu",
+                         static_cast<unsigned long long>(st.st_ino));
 }
 #endif
